Выкинул size_t так как %zu не переносимо, а искать замену лень

+++ PnP_convMain.cpp
@@ -868,11 +868,11 @@ void PnP_convFrame::On_mnuSaveProdSelected(wxCommandEvent& event)
 	wxXmlNode *templates_node = new wxXmlNode(root_node, wxXML_ELEMENT_NODE, "Templates");
 
 	size_t comp_count = m_components_list.GetCount();
-	for (size_t subpcb_index = 0; (long)subpcb_index < subpcbs; subpcb_index++)
+	for (int subpcb_index = 0; subpcb_index < subpcbs; subpcb_index++)
 	{
 		wxXmlNode *panel_node = new wxXmlNode(wxXML_ELEMENT_NODE, "Panel");
 		panels_node->AddChild(panel_node);
-		panel_node->AddAttribute("Ref", wxString::Format("%zu", subpcb_index+1));
+		panel_node->AddAttribute("Ref", wxString::Format("%d", subpcb_index+1));
 		panel_node->AddAttribute("Template", m_project.pcbs[subpcb_index].subpcb_name);
 		panel_node->AddAttribute("Position", wxString::Format("%ld,%ld", (long)((m_project.pcbs[subpcb_index].offset_x-m_project.offset_x)*1000),
 										 (long)((m_project.pcbs[subpcb_index].offset_y-m_project.offset_y)*1000)));
